CNV-40846: reorganize utils folder
