+++ src/utils/constants/ui.ts
@@ -1,0 +1,2 @@
+export const SHARED_DEFAULT_PATH_NEW_RESOURCE_YAML = '~new';
+export const SHARED_DEFAULT_PATH_NEW_RESOURCE_FORM = '~new/form';
