+++ src/utils/utils/ui.ts
@@ -1,0 +1,2 @@
+export const getContentScrollableElement = (): HTMLElement =>
+  document.getElementById('content-scrollable');
