+++ src/views/nads/list/NetworkAttachmentDefinitionList.tsx
@@ -14,19 +14,19 @@ import {
   VirtualizedTable,
 } from '@openshift-console/dynamic-plugin-sdk';
 import ListEmptyState from '@utils/components/ListEmptyState/ListEmptyState';
-import { useNetworkingTranslation } from '@utils/hooks/useNetworkingTranslation';
-import { NetworkAttachmentDefinitionKind } from '@utils/resources/nads/types';
+import { DEFAULT_NAMESPACE } from '@utils/constants';
 import {
-  resourcePathFromModel,
   SHARED_DEFAULT_PATH_NEW_RESOURCE_FORM,
   SHARED_DEFAULT_PATH_NEW_RESOURCE_YAML,
-} from '@utils/utils';
-import { isEmpty } from '@utils/utils/utils';
+} from '@utils/constants/ui';
+import { useNetworkingTranslation } from '@utils/hooks/useNetworkingTranslation';
+import { NetworkAttachmentDefinitionKind } from '@utils/resources/nads/types';
+import { resourcePathFromModel } from '@utils/resources/shared';
+import { isEmpty } from '@utils/utils';
 
 import NADListEmpty from './components/NADListEmpty/NADListEmpty';
 import NADsRow from './components/NADsRow/NADsRow';
 import useNADsColumns from './hooks/useNADsColumns';
-import { DEFAULT_NAMESPACE } from '@utils/constants';
 
 type NetworkAttachmentDefinitionListProps = {
   namespace: string;
